refactor(use-cases): mix use cases for personas in one file
