--- app/src/main/java/com/anthonychaufrias/people/domain/UpdPersonasUseCase.kt
@@ -1,10 +1,0 @@
-package com.anthonychaufrias.people.domain
-
-import com.anthonychaufrias.people.data.PersonaRepository
-import com.anthonychaufrias.people.data.model.Persona
-
-class UpdPersonasUseCase {
-    private val repository = PersonaRepository()
-
-    suspend operator fun invoke(persona: Persona): Persona = repository.updatePersona(persona)
-}
